requirements.txt
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
XlsxWriter==3.0.3